+++ patientservice/src/main/java/com/medilabosolutions/type2diabetesfinder/patientservice/repository/PatientRepository.java
@@ -4,6 +4,6 @@ import com.medilabosolutions.type2diabetesfinder.patientservice.model.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PatientRepository extends JpaRepository<Patient, Long>, JpaSpecificationExecutor<Patient> {
+public interface PatientRepository extends JpaRepository<Patient, Integer>, JpaSpecificationExecutor<Patient> {
 
 }
